Petite modif

+++ src/Controller/DeliveryPaymentController.php
@@ -24,40 +24,45 @@ class DeliveryPaymentController extends AbstractController
         $form->handleRequest($request);
         
         if ($form->isSubmitted() && $form->isValid()){
-            //Date du jour de la commande
-            $order->setDate(new \Datetime());
-            //Qui à fait la commande (celui qui est connecter)
-            $order->setUser($this->getUser());
-            //Récupere le panier dans la session
-            $session = $this->requestStack->getSession();
-            $panier = $session->get('panier');
-            //Récupere le prix total de la commande
-            $order->setTotal($cartService->getTotal());
-            //Enregistre la commande
-            $entityManager->persist($order);
-            
-            //parcour le panier
-            foreach($panier as $key => $value){
-                //Création d'un nouveau objet orderProduct (nouvelle ligne dans orderProduct)
-                $orderProduct = new OrderProduct();
-                //Lie cette ligne à la commande
-                $order->addOrderProduct($orderProduct);
-                //Récupère le produit a partir de l'id
-                $produit = $productRepository->find($key);
-                //Lie le produit à la commande
-                $orderProduct->setProduct($produit);
-                //On lui donne la quantité
-                $orderProduct->setQuantity($value);
-                //Enregistrement
-                $entityManager->persist($orderProduct);
+            if ($order->getDelivery() === null ){
+                $this->addFlash('formIsNotValid', 'Veuillez renseigner l\'adresse de livraison');
+                return $this->redirectToRoute("delivery_payment");
+            } else {
+                //Date du jour de la commande
+                $order->setDate(new \Datetime());
+                //Qui à fait la commande (celui qui est connecter)
+                $order->setUser($this->getUser());
+                //Récupere le panier dans la session
+                $session = $this->requestStack->getSession();
+                $panier = $session->get('panier');
+                //Récupere le prix total de la commande
+                $order->setTotal($cartService->getTotal());
+                //Enregistre la commande
+                $entityManager->persist($order);
                 
-                dump($produit);
+                //parcour le panier
+                foreach($panier as $key => $value){
+                    //Création d'un nouveau objet orderProduct (nouvelle ligne dans orderProduct)
+                    $orderProduct = new OrderProduct();
+                    //Lie cette ligne à la commande
+                    $order->addOrderProduct($orderProduct);
+                    //Récupère le produit a partir de l'id
+                    $produit = $productRepository->find($key);
+                    //Lie le produit à la commande
+                    $orderProduct->setProduct($produit);
+                    //On lui donne la quantité
+                    $orderProduct->setQuantity($value);
+                    //Enregistrement
+                    $entityManager->persist($orderProduct);
+                    
+                    dump($produit);
+                }
+                //Exécution des requètes
+                $entityManager->flush();
+                //-----------------
+                $cartService->removeAll();
+                return $this->redirectToRoute("order_validated");
             }
-            //Exécution des requètes
-            $entityManager->flush();
-            //-----------------
-            $cartService->removeAll();
-            return $this->redirectToRoute("order_validated");
         }
         
         return $this->render('order_tunnel/deliveryPayment.html.twig', [
